Refactoring code according to code review
